# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Netumo Clone API",
        "description": "API documentation for uptime monitoring system",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
        "/api/targets": {
            "get": {
                "tags": [
                    "Targets"
                ],
                "summary": "List all targets",
                "operationId": "540c990d268cf9c35b2785098c1979dc",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TargetResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Targets"
                ],
                "summary": "Create a new target",
                "operationId": "cce13a0b839fce00e4385bcb42456fa6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TargetResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Target created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TargetResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/targets/{id}": {
            "get": {
                "tags": [
                    "Targets"
                ],
                "summary": "Get target by ID",
                "operationId": "02d9296809be9e9e55d2f8ac32c3ab7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TargetResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Target not found"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Targets"
                ],
                "summary": "Update target",
                "operationId": "32d2541777861ec66396a8bbe8a2bcda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TargetResource"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Target updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TargetResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Target not found"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Targets"
                ],
                "summary": "Delete target",
                "operationId": "db26aefe618fc4b148e608fa1ca15e61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Target not found"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "TargetResource": {
                "title": "Target Resource",
                "description": "Target resource representation",
                "properties": {
                    "id": {
                        "description": "Target unique identifier",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Target name",
                        "type": "string",
                        "example": "My Website"
                    },
                    "url": {
                        "description": "Target URL to monitor",
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com"
                    },
                    "created_at": {
                        "description": "Target creation date and time",
                        "type": "string",
                        "format": "datetime",
                        "example": "Mon 04 Jun, 2025 20:30"
                    }
                },
                "type": "object"
            },
            "OptionResource": {
                "description": "A trait that provides JSON formatting options for API resources",
                "properties": {
                    "": {
                        "description": "JSON encoding options value",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "UnauthorizedError": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Targets",
            "description": "Targets"
        }
    ]
}